feat(testing): Move generic matchers to utils for later reuse
