+++ dd-java-agent/instrumentation/opentelemetry/opentelemetry-1.4/src/test/java/opentelemetry14/OpenTelemetry14ConventionsTest.java
@@ -1,6 +1,5 @@
 package opentelemetry14;
 
-import static datadog.trace.agent.test.assertions.Matchers.is;
 import static datadog.trace.agent.test.assertions.SpanMatcher.span;
 import static datadog.trace.agent.test.assertions.TagsMatcher.defaultTags;
 import static datadog.trace.agent.test.assertions.TagsMatcher.tag;
@@ -9,6 +8,7 @@ import static datadog.trace.api.DDTags.ANALYTICS_SAMPLE_RATE;
 import static datadog.trace.api.DDTags.DD_SVC_SRC;
 import static datadog.trace.bootstrap.instrumentation.api.Tags.HTTP_STATUS;
 import static datadog.trace.bootstrap.instrumentation.api.Tags.SPAN_KIND;
+import static datadog.trace.junit.utils.assertions.Matchers.is;
 import static io.opentelemetry.api.common.AttributeKey.longKey;
 import static io.opentelemetry.api.common.AttributeKey.stringKey;
 import static io.opentelemetry.api.trace.SpanKind.CLIENT;
@@ -20,10 +20,10 @@ import static java.util.Collections.emptyMap;
 import static org.junit.jupiter.params.provider.Arguments.arguments;
 
 import datadog.opentelemetry.shim.trace.OtelConventions;
-import datadog.trace.agent.test.assertions.Matcher;
-import datadog.trace.agent.test.assertions.Matchers;
 import datadog.trace.agent.test.assertions.TagsMatcher;
 import datadog.trace.bootstrap.instrumentation.api.ServiceNameSources;
+import datadog.trace.junit.utils.assertions.Matcher;
+import datadog.trace.junit.utils.assertions.Matchers;
 import io.opentelemetry.api.trace.Span;
 import io.opentelemetry.api.trace.SpanBuilder;
 import io.opentelemetry.api.trace.SpanKind;
