+++ utils/test-junit-utils/src/main/java/datadog/trace/junit/utils/assertions/Any.java
@@ -1,4 +1,4 @@
-package datadog.trace.agent.test.assertions;
+package datadog.trace.junit.utils.assertions;
 
 import java.util.Optional;
