+++ utils/test-junit-utils/src/main/java/datadog/trace/junit/utils/assertions/IsFalse.java
@@ -1,4 +1,4 @@
-package datadog.trace.agent.test.assertions;
+package datadog.trace.junit.utils.assertions;
 
 import java.util.Optional;
